Done debugging MCMC and SimAnneal for new RAM saver updates

+++ code/pythonSide/SimulationManager.py
@@ -384,7 +384,7 @@ def simulator(paramSettingsDict):
     ############################################################
     # finish the Gelman-Rubin statistic calculations if requested
     ############################################################
-    if (paramSettingsDict['CalcGelmanRubin']and paramSettingsDict['useMultiProcessing'])and(paramSettingsDict["mcONLY"]==False):
+    if (paramSettingsDict['CalcGelmanRubin']and paramSettingsDict['useMultiProcessing'])and((paramSettingsDict["mcONLY"]==False)and(paramSettingsDict['simAnneal']==False)):
         tools.gen.gelmanRubinStage2(dataFiles)    
     
     ####################################################################################################
@@ -437,7 +437,7 @@ def simulator(paramSettingsDict):
                 origFiles.append(os.path.basename(MCMCsummaryRootPlotFilename)+"-chain_"+str(chainNum)+".png" )
                 if paramSettingsDict['DIonly']==False:
                     origFiles.append(os.path.basename(MCMCsummaryRootPlotFilename)+"-chain_"+str(chainNum)+"-RVoffsets.png")
-    if paramSettingsDict['makeMCMCprogPlots']:
+    elif paramSettingsDict['makeMCMCprogPlots']:
         for chainNum in range(1,len(dataFiles)+1):
             origFiles.append(os.path.basename(MCMCsummaryRootPlotFilename)+"-chain_"+str(chainNum)+".png")
             if paramSettingsDict['DIonly']==False:
@@ -448,7 +448,7 @@ def simulator(paramSettingsDict):
                 if paramSettingsDict['DIonly']==False:
                     origFiles.append(os.path.basename(simAnnealSummaryRootPlotFilename)+"-chain_"+str(chainNum)+"-RVoffsets.png")
     # get GR filenames
-    if (paramSettingsDict['CalcGelmanRubin']and paramSettingsDict['useMultiProcessing'])and(paramSettingsDict["mcONLY"]==False):
+    if (paramSettingsDict['CalcGelmanRubin']and paramSettingsDict['useMultiProcessing'])and((paramSettingsDict["mcONLY"]==False)and(paramSettingsDict['simAnneal']==False)):
         origFiles.append('GRvalues.txt')
         origFiles.append('Tvalues.txt')
     # get RESULTS.txt filenames
@@ -514,14 +514,14 @@ def simulator(paramSettingsDict):
             for filename in dataFiles:
                 print 'Deleting file: '+os.path.basename(filename)
                 os.remove(filename)
-            if (paramSettingsDict['removeBurnIn'] and (paramSettingsDict["mcONLY"]==False))and (paramSettingsDict['CalcBurnIn']and (paramSettingsDict['simAnneal']==False)):
+            if (paramSettingsDict['removeBurnIn'] and (paramSettingsDict["mcONLY"]==False))and(paramSettingsDict['CalcBurnIn']and (paramSettingsDict['simAnneal']==False)):
                 print '\nDeleting Burn-In removed MCMC chain data files\n'+"-"*40
                 for filename in strippedNames:
                     print 'Deleting file: '+os.path.basename(filename)
                     os.remove(filename)
             
     ## delete GR chain files if requested
-    if (paramSettingsDict['CalcGelmanRubin'] and(paramSettingsDict['delGRchainFiles'] and (len(dataFiles)>1)))and(paramSettingsDict["mcONLY"]==False):
+    if (paramSettingsDict['CalcGelmanRubin'] and(paramSettingsDict['delGRchainFiles'] and (len(dataFiles)>1)))and((paramSettingsDict["mcONLY"]==False)and (paramSettingsDict['simAnneal']==False)):
         print '\n\nDeleting final output GR chain value files\n'+"-"*40
         for chainNum in range(1,len(dataFiles)+1):
             filename = os.path.join(paramSettingsDict['outputData_dir'],"gelmanRubin-chain_"+str(chainNum)+".txt")
@@ -535,7 +535,7 @@ def simulator(paramSettingsDict):
         if cleanDataFilename!='':
             print 'Deleting file: '+cleanDataFilename
             os.remove(cleanDataFilename)
-        if (paramSettingsDict['removeBurnIn'] and (paramSettingsDict["mcONLY"]==False))and paramSettingsDict['CalcBurnIn']:
+        if (paramSettingsDict['removeBurnIn'] and (paramSettingsDict["mcONLY"]==False))and (paramSettingsDict['CalcBurnIn']and (paramSettingsDict['simAnneal']==False)):
             print 'Deleting file: '+dataFinalFilename2
             os.remove(dataFinalFilename2)
